used enum class indexStress2DInterface

+++ applications/GeoMechanicsApplication/custom_constitutive/small_strain_udsm_3D_interface_law.cpp
@@ -31,7 +31,7 @@ std::size_t GetIndex3D(const indexStress3DInterface Index3D)
     case INDEX_3D_INTERFACE_XZ:
         return static_cast<std::size_t>(indexStress3D::INDEX_3D_XZ);
     default:
-        KRATOS_ERROR << "invalid index: " << Index3D << std::endl;
+        KRATOS_ERROR << "invalid index: " << static_cast<int>(Index3D) << std::endl;
     }
 }
